+++ CodeStarterCamp_Week1/main.swift
@@ -1,51 +1,10 @@
 //
-//  main.swift
+//  week1_step3_tag.swift
 //  CodeStarterCamp_Week1
 //
-//  Created by yagom.
-//  Copyright © yagom academy. All rights reserved.
+//  Created by 윤해수 on 2023/02/23.
 //
-//  CodeStarterCamp_Week1_tag
 
 import Foundation
 
-print("Hello, World!")
-
-func icecream() {
-    print("***********")
-}
-
-func stick() {
-    print("    | |")
-}
-
-for _ in 1...8 {
-    icecream()
-}
-
-for _ in 1...4 {
-    stick()
-}
-
-//var icecreamwidth = 11
-//var icecreamlength = 8
-////var stickwidth = 1 *막대기 사이의 공간도 조절할 수 있을까?
-//var sticklength = 4
-//
-//func icecream() {
-//    for _ in 1...icecreamlength {
-//        for _ in 1...icecreamwidth {
-//            print("*", terminator: "")
-//        }
-//        print("")
-//    }
-//}
-//
-//func stick() {
-//    for _ in 1...sticklength {
-//        print("    | |")
-//        }
-//    }
-//
-//icecream()
-//stick()
+print("<정보>\n길이:\n몸통:\n토핑:\n막대길이:")
