+++ CodeStarterCamp_Week1/week1_step2.swift
@@ -1,0 +1,51 @@
+//
+//  main.swift
+//  CodeStarterCamp_Week1
+//
+//  Created by yagom.
+//  Copyright © yagom academy. All rights reserved.
+//
+//  CodeStarterCamp_Week1_tag
+
+import Foundation
+
+print("Hello, World!")
+
+func icecream() {
+    print("***********")
+}
+
+func stick() {
+    print("    | |")
+}
+
+for _ in 1...8 {
+    icecream()
+}
+
+for _ in 1...4 {
+    stick()
+}
+
+//var icecreamwidth = 11
+//var icecreamlength = 8
+////var stickwidth = 1 *막대기 사이의 공간도 조절할 수 있을까?
+//var sticklength = 4
+//
+//func icecream() {
+//    for _ in 1...icecreamlength {
+//        for _ in 1...icecreamwidth {
+//            print("*", terminator: "")
+//        }
+//        print("")
+//    }
+//}
+//
+//func stick() {
+//    for _ in 1...sticklength {
+//        print("    | |")
+//        }
+//    }
+//
+//icecream()
+//stick()
