--- CodeStarterCamp_Week1/week1_step3_tag.swift
@@ -1,8 +1,0 @@
-//
-//  week1_step3_tag.swift
-//  CodeStarterCamp_Week1
-//
-//  Created by 윤해수 on 2023/02/23.
-//
-
-import Foundation
